masim-record: Increase acceptable error for 2mb pattern
On my new machine, masim-record test fails for 2mb pattern, since the reported
WSS is larger than the maximum value we expect.  The WSS on the machine is
consistent, though.  It is about 9-10 MiB.  It seems not that bad.  Increase
acceptable error rate of the test to fit with it.

Signed-off-by: SeongJae Park <sj@kernel.org>